subscriber add and manage

+++ app/Http/Controllers/Author/PostController.php
@@ -102,6 +102,11 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         return view('author.post.show',compact('post'));
     }
 
@@ -113,6 +118,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit',compact('post','categories','tags'));
@@ -127,6 +137,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         $this->validate($request,[
             'title' => 'required',
             'image' => 'image',
@@ -187,6 +202,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if ($post->user_id != Auth::id())
+        {
+            Toastr::error('You are not authorized to access this post','Error');
+            return redirect()->back();
+        }
         if (Storage::disk('public')->exists('post/'.$post->image))
         {
             Storage::disk('public')->delete('post/'.$post->image);
